django/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor-5==0.2.15
django-filebrowser==4.0.3
django-grappelli==3.0.9
django-js-asset==2.2.0
pillow==11.0.0
psycopg2-binary==2.9.10
six==1.16.0
sqlparse==0.5.1
